[clang] Move more function from `Sema` to `SemaObjC`
